requirements.txt
Django>=4.1,<4.2
wagtail>=4.2,<4.3
opencv-python>=4.7.0
djangorestframework>=3.14
django-cors-headers>=3.14

